Move resource registration to centralized spot
