+++ src/engine/renderer/renderer.main.ts
@@ -2,30 +2,7 @@ import { defineModule, Thread } from "../module/module.common";
 import { IMainThreadContext } from "../MainThread";
 import { WorkerMessageType } from "../WorkerMessage";
 import { RendererMessageType, rendererModuleName } from "./renderer.common";
-import { registerResource } from "../resource/resource.main";
 import { createDisposables } from "../utils/createDisposables";
-import {
-  AccessorResource,
-  BufferResource,
-  BufferViewResource,
-  CameraResource,
-  ImageResource,
-  InstancedMeshResource,
-  InteractableResource,
-  LightMapResource,
-  LightResource,
-  MaterialResource,
-  MeshPrimitiveResource,
-  MeshResource,
-  ReflectionProbeResource,
-  SamplerResource,
-  SkinResource,
-  TextureResource,
-  TilesRendererResource,
-} from "../resource/schema";
-import { MainNode } from "../node/node.main";
-import { MainThreadNametagResource } from "../nametag/nametag.main";
-import { MainScene } from "../scene/scene.main";
 
 type MainRendererModuleState = {};
 
@@ -57,31 +34,7 @@ export const RendererModule = defineModule<IMainThreadContext, MainRendererModul
       canvasWidth: ctx.canvas.clientWidth,
       canvasHeight: ctx.canvas.clientHeight,
     });
-    return createDisposables([
-      registerResizeEventHandler(ctx),
-      registerResource(ctx, LightResource),
-      registerResource(ctx, SamplerResource),
-      registerResource(ctx, CameraResource),
-      registerResource(ctx, BufferResource),
-      registerResource(ctx, BufferViewResource),
-      registerResource(ctx, ImageResource),
-      registerResource(ctx, MaterialResource),
-      registerResource(ctx, TextureResource),
-      registerResource(ctx, MeshResource),
-      registerResource(ctx, MainNode),
-      registerResource(ctx, MainScene),
-      registerResource(ctx, MainThreadNametagResource),
-      registerResource(ctx, MeshPrimitiveResource),
-      registerResource(ctx, InteractableResource),
-      registerResource(ctx, AccessorResource),
-      registerResource(ctx, MeshResource),
-      registerResource(ctx, MeshPrimitiveResource),
-      registerResource(ctx, SkinResource),
-      registerResource(ctx, InstancedMeshResource),
-      registerResource(ctx, LightMapResource),
-      registerResource(ctx, ReflectionProbeResource),
-      registerResource(ctx, TilesRendererResource),
-    ]);
+    return createDisposables([registerResizeEventHandler(ctx)]);
   },
 });
